+++ CHANGES.md
@@ -29,6 +29,7 @@ Changes in 11.0
 - SQL add ORDER BY to GetChildrenMP() in GetMP.php
 - SQL select Grading Scales by Teacher, only the ones having student grades in GradeBreakdown.php
 - Add Insert into or Update DB functions in DBUpsert.php
+- Use DBUpsert() function in Template.fnc.php, Config.fnc.php & StudentAssignments.fnc.php
 
 Changes in 10.9.6
 -----------------